WIP remove more unused code and refactor
