+++ app/routers/verifier.py
@@ -1,16 +1,15 @@
 import logging
 import time
 import traceback
-from typing import List, Optional
+from typing import List
 from dependencies import *
 from facade import (
-    # create_controller,
     get_schema_attributes,
     send_proof_request,
     verify_proof_req,
 )
 from aries_cloudcontroller import AriesAgentControllerBase
-from fastapi import APIRouter, Header, HTTPException, Query, Depends
+from fastapi import APIRouter, HTTPException, Query, Depends
 from utils import construct_indy_proof_request, construct_zkp
 from schemas import RequestProofResponse
 
@@ -31,9 +30,6 @@ async def get_proof_request(
     self_attested: List[str] = None,
     revocation: int = None,
     exchange_tracing: bool = False,
-    api_key: Optional[str] = Header(None),
-    wallet_id: Optional[str] = Header(None),
-    tenant_jwt: Optional[str] = Header(None),
     aries_controller: AriesAgentControllerBase = Depends(yoma_agent),
 ):
     """
@@ -60,13 +56,8 @@ async def get_proof_request(
     presentation_exchange_id: json
         The presentation exchange ID JSON object
     """
-    auth_headers = {
-        "api_key": api_key,
-        "wallet_id": wallet_id,
-        "tenant_jwt": tenant_jwt,
-    }
     try:
-        # async with create_controller(auth_headers) as controller:
+        # We can refactor this - it's already context managed
         async with aries_controller as controller:
             schema_resp = await get_schema_attributes(controller, schema_id)
             is_attrs_match = all(x in schema_resp for x in requested_attrs)
@@ -124,9 +115,7 @@ async def get_proof_request(
 @router.get("/verify-proof-request", tags=["verifier", "proof"])
 async def verify_proof_request(
     presentation_exchange_id: str,
-    api_key: Optional[str] = Header(None),
-    wallet_id: Optional[str] = Header(None),
-    tenant_jwt: Optional[str] = Header(None),
+    aries_controller: AriesAgentControllerBase = Depends(member_agent),
 ):
     """
     Verify a proof request against the ledger
@@ -146,13 +135,8 @@ async def verify_proof_request(
     verify: dict
         The json representation of the verify request
     """
-    auth_headers = {
-        "api_key": api_key,
-        "wallet_id": wallet_id,
-        "tenant_jwt": tenant_jwt,
-    }
     try:
-        async with create_controller(auth_headers) as controller:
+        async with aries_controller as controller:
             verify = await verify_proof_req(controller, presentation_exchange_id)
 
             if not verify["state"] == "verified":
