+++ app/tests/conftest.py
@@ -5,7 +5,6 @@ from aries_cloudcontroller import AriesAgentControllerBase
 from aries_cloudcontroller.controllers.ledger import LedgerController
 from aries_cloudcontroller.controllers.wallet import WalletController
 from mockito import mock
-from fastapi import Depends
 
 import ledger_facade
 import utils
