+++ junit.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="30" time="11.209" timestamp="2021-06-23T16:00:40.075734" hostname="localhost"><testcase classname="tests.test_acapy_ledger_facade" name="test_error_on_get_taa" time="0.007" /><testcase classname="tests.test_acapy_ledger_facade" name="test_error_on_accept_taa" time="0.004" /><testcase classname="tests.test_acapy_ledger_facade" name="test_error_on_get_did_endpoint" time="0.005" /><testcase classname="tests.test_acapy_wallet_facade" name="test_error_on_get_pub_did" time="0.005" /><testcase classname="tests.test_acapy_wallet_facade" name="test_error_on_assign_pub_did" time="0.004" /><testcase classname="tests.test_acapy_wallet_facade" name="test_error_on_create_pub_did" time="0.005" /><testcase classname="tests.test_facade" name="test_create_controller[fake_header0-admin]" time="0.005" /><testcase classname="tests.test_facade" name="test_create_controller[fake_header1-tenant]" time="0.003" /><testcase classname="tests.test_facade" name="test_create_controller[fake_header2-admin]" time="0.004" /><testcase classname="tests.test_facade" name="test_create_controller[fake_header3-admin]" time="0.004" /><testcase classname="tests.test_facade" name="test_create_controller[fake_header4-False]" time="0.004" /><testcase classname="tests.test_facade" name="test_create_controller[fake_header5-False]" time="0.003" /><testcase classname="tests.test_utils" name="test_get_controller_type[fake_header0-admin]" time="0.003" /><testcase classname="tests.test_utils" name="test_get_controller_type[fake_header1-tenant]" time="0.003" /><testcase classname="tests.test_utils" name="test_get_controller_type[fake_header2-admin]" time="0.003" /><testcase classname="tests.test_utils" name="test_get_controller_type[fake_header3-admin]" time="0.003" /><testcase classname="tests.test_utils" name="test_get_controller_type[fake_header4-None]" time="0.003" /><testcase classname="tests.test_utils" name="test_get_controller_type[fake_header5-None]" time="0.003" /><testcase classname="tests.test_utils" name="test_controller_factory[fake_header0-ABCMeta]" time="0.004" /><testcase classname="tests.test_utils" name="test_controller_factory[fake_header1-ABCMeta]" time="0.004" /><testcase classname="tests.test_utils" name="test_controller_factory[fake_header2-ABCMeta]" time="0.004" /><testcase classname="tests.test_utils" name="test_controller_factory[fake_header3-ABCMeta]" time="0.004" /><testcase classname="tests.test_utils" name="test_controller_factory[fake_header4-False]" time="0.003" /><testcase classname="tests.test_utils" name="test_controller_factory[fake_header5-False]" time="0.003" /><testcase classname="tests.test_utils" name="test_construct_zkp" time="0.001" /><testcase classname="tests.test_utils" name="test_construct_zkp_empty" time="0.001" /><testcase classname="tests.test_utils" name="test_construct_indy_proof_request" time="0.001" /><testcase classname="tests.test_via_web_wallet" name="test_get_pub_did_via_web" time="3.940" /><testcase classname="tests.test_wallet" name="test_create_public_did" time="5.960" /><testcase classname="tests.test_wallet" name="test_create_public_did_no_api_key" time="0.005" /></testsuite></testsuites>
+<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="3" time="0.066" timestamp="2021-07-02T12:20:43.709529" hostname="localhost"><testcase classname="tests.test_acapy_ledger_facade" name="test_error_on_get_taa" time="0.006" /><testcase classname="tests.test_acapy_ledger_facade" name="test_error_on_accept_taa" time="0.003" /><testcase classname="tests.test_acapy_ledger_facade" name="test_error_on_get_did_endpoint" time="0.003" /></testsuite></testsuites>
